requirements.txt
torch>=2.0
transformers>=4.51
Pillow>=9.0
datasets>=2.0
